requirements.txt
gunicorn==19.6.0
Django==1.10
django-import-export==0.5.1hero